requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
transformers==4.35.0
torch==2.0.1
pillow==9.5.0
python-multipart==0.0.6
jinja2==3.1.2
pydantic-settings==2.0.3
pydantic==2.5.0
numpy==1.24.3
requests==2.31.0
accelerate==0.24.0
timm==0.9.2
easyocr==1.7.1
opencv-python==4.8.1.78
